+++ internal/storage/ydb_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-
 	"github.com/ydb-platform/ydb-go-sdk/v3/table/types"
 )