+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Changed template of static credentials config option
+* Fixed the initialization of static credentials for secure/insecure endpoints 
+
 ## v1.0.0
 
 ### Added